V1.0

+++ src/components/DeleteArticle.jsx
@@ -6,23 +6,26 @@ import { deleteObject, ref } from 'firebase/storage'
 
 export const DeleteArticle = ({ id, imageUrl }) => {
     const handleDelete = async () => {
-        try {
-            await deleteDoc(doc(db, 'Articles', id))
-            toast(`Article deleted successfully (blog id:${id})`)
-            const storageRef = ref(storage, imageUrl)
-            await deleteObject(storageRef)
-        } catch (error) {
-            toast('error deleting article', { type: 'error' })
-            console.log(error)
+        if (window.confirm('Are you sure to delete your blog?')) {
+            try {
+                await deleteDoc(doc(db, 'Articles', id))
+                toast(`Article deleted successfully (blog id:${id})`)
+                const storageRef = ref(storage, imageUrl)
+                await deleteObject(storageRef)
+            } catch (error) {
+                toast('error deleting article', { type: 'error' })
+                console.log(error)
+            }
         }
     }
 
     return (
         <div>
-            <button
-                className='btn btn-danger'
+            <i
+                className='fa fa-times'
+                style={{ cursor: 'pointer' }}
                 onClick={handleDelete}
-            >Delete</button>
+            />
         </div>
     )
 }
